ensure if a filter is in (or not in) a list before removing/adding

+++ cogs/filter_commands.py
@@ -28,6 +28,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
 
         # Opens a connection and inerts the text filter into the textfilters database
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM textfilters WHERE userid == $1 AND textfilter == $2", ctx.author.id, filter)
+
+            if curr_filter_rows: # Text filter is already in list
+                return await ctx.send((f"{filter} is already in your filters list!"))
+
             await db("INSERT INTO textfilters (userid, textfilter) VALUES ($1, $2);", ctx.author.id, filter)
 
         await ctx.send(f"Added `{filter}` to your text filter list", allowed_mentions=discord.AllowedMentions(everyone=False, users=False, roles=False))
@@ -40,6 +45,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
         """
 
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM channelfilters WHERE userid == $1 AND channelfilter == $2", ctx.author.id, filter)
+
+            if curr_filter_rows: # Channel filter is already in list
+                return await ctx.send((f"{filter} is already in your filters list!"))
+
             await db("INSERT INTO channelfilters (userid, channelfilter) VALUES ($1, $2);", ctx.author.id, filter.id)
         await ctx.send(f"Added {filter.mention} to your channel filter list")
 
@@ -62,6 +72,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
 
         # Opens a connection and inerts the server filter into the serverfilters database
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM serverfilters WHERE userid == $1 AND serverfilter == $2", ctx.author.id, filter)
+
+            if curr_filter_rows: # Server filter is already in list
+                return await ctx.send((f"{filter} is already in your filters list!"))\
+
             await db("INSERT INTO serverfilters (userid, serverfilter) VALUES ($1, $2);", ctx.author.id, filter)
 
         await ctx.send(f"Added `{filter}` to your server filter list")
@@ -72,6 +87,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
 
         # Opens a connection and inerts the user filter into the serverfilters database
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM userfilters WHERE userid == $1 AND userfilter == $2", ctx.author.id, filter)
+
+            if curr_filter_rows: # User filter is already in list
+                return await ctx.send((f"{filter} is already in your filters list!"))
+
             await db("INSERT INTO userfilters (userid, userfilter) VALUES ($1, $2);", ctx.author.id, filter)
 
         await ctx.send(f"Added `{filter}` to your user filter list", allowed_mentions=discord.AllowedMentions(everyone=False, users=False, roles=False))
@@ -126,6 +146,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
         """Removes a text filter"""
 
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM textfilters WHERE userid == $1 AND textfilter == $2", ctx.author.id, filter)
+
+            if not curr_filter_rows: # Text filter is already not in list
+                return await ctx.send((f"{filter} is not a filter in your list!"))
+
             await db("DELETE FROM textfilters WHERE userid=$1 and textfilter=$2;", ctx.author.id, filter)
         return await ctx.send(f"Removed `{filter}` from your text filter list")
 
@@ -135,6 +160,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
         """Removes a channel filter"""
 
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM channelfilters WHERE userid == $1 AND channelfilter == $2", ctx.author.id, filter)
+
+            if not curr_filter_rows: # Channel filter is already not in list
+                return await ctx.send((f"{filter} is not a filter in your list!"))
+
             await db("DELETE FROM channelfilters WHERE userid=$1 AND channelfilter=$2;", ctx.author.id, filter.id)
         return await ctx.send(f"Removed `{filter.mention}` from your channel filter list")
 
@@ -152,6 +182,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
 
         # Opens a connection and inerts the text filter into the serverfilters database
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM serverfilters WHERE userid == $1 AND serverfilter == $2", ctx.author.id, filter)
+
+            if not curr_filter_rows: # Server filter is already not in list
+                return await ctx.send((f"{filter} is not a filter in your list!"))
+
             await db("DELETE FROM serverfilters WHERE userid=$1 AND serverfilter=$2;", ctx.author.id, filter)
 
         await ctx.send(f"You will now get message triggers from `{filter}`")
@@ -162,6 +197,11 @@ class FilterCommands(vbu.Cog, name="Filter Commands"):
 
         # Opens a connection and inerts the user filter into the userfilters database
         async with vbu.Database() as db:
+            curr_filter_rows = await db("SELECT FROM userfilters WHERE userid == $1 AND userfilter == $2", ctx.author.id, filter)
+
+            if not curr_filter_rows: # User filter is already not in list
+                return await ctx.send((f"{filter} is not a filter in your list!"))
+
             await db("DELETE FROM userfilters WHERE userid=$1 AND userfilter=$2;", ctx.author.id, filter)
 
         await ctx.send(f"You will now recieve message triggers from `{filter}`.")
